backend/requirements.txt
amqp==2.4.2
appnope==0.1.0
asgiref==2.3.2
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.2.1
Automat==0.7.0
backcall==0.1.0
billiard==3.5.0.5
celery==4.2.1
channels==2.1.7
constantly==15.1.0
daphne==2.2.5
decorator==4.4.0
Django==2.1.7
django-cors-headers==2.5.2
djangorestframework==3.9.2
djangorestframework-simplejwt==4.1.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
ipython==7.3.0
ipython-genutils==0.2.0
jedi==0.13.3
kombu==4.4.0
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
prompt-toolkit==2.0.9
psycopg2-binary==2.7.7
ptyprocess==0.6.0
Pygments==2.3.1
PyHamcrest==1.9.0
PyJWT==1.7.1
python-decouple==3.1
pytz==2018.9
six==1.12.0
traitlets==4.3.2
Twisted==18.9.0
txaio==18.8.1
vine==1.2.0
wcwidth==0.1.7
zope.interface==4.6.0
